+++ didelphis-common-io/src/main/java/org/didelphis/io/DiskFileHandler.java
@@ -23,8 +23,6 @@ import lombok.EqualsAndHashCode;
 import lombok.NonNull;
 import lombok.ToString;
 
-import org.didelphis.utilities.Logger;
-
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileInputStream;
@@ -44,8 +42,6 @@ import java.io.Writer;
 @ToString
 @EqualsAndHashCode
 public final class DiskFileHandler implements FileHandler {
-
-	private static final Logger LOG = Logger.create(DiskFileHandler.class);
 
 	private final String encoding;